extra feature update

+++ scripts/tiles/barrel.py
@@ -4,13 +4,37 @@ from ..particle import Particle
 
 
 class Barrel(Tile):
+    """
+    A barrel tile. It can explode and create a smoke explosion that kills the player and enemies around it.
+    """
+
     def __init__(self, image, pos, variant, size, smoke_animation):
+        """
+        Create a new Barrel object.
+
+        Parameters:
+            image (pygame.Surface): The image of the barrel.
+            pos (tuple[int, int]): The position of the barrel.
+            variant (int): The variant of the barrel.
+            size (int): The size of the barrel.
+            smoke_animation (Animation): The animation of the smoke particles that will appear when the barrel explodes. It should be an Animation object.
+        """
+
         super().__init__(image, pos, "barrel", variant, size, False)
         self.smoke_explosion = None
         self.exploded = False
         self.smoke_animation = smoke_animation
+        self.killed = False
 
     def render(self, surf, offset):
+        """
+        Render the barrel on the screen.
+
+        Parameters:
+            surf (pygame.Surface): The surface to render the barrel.
+            offset (tuple[int, int]): The offset of the screen, used to render the barrel in the correct position.
+        """
+
         if not self.exploded:
             super().render(surf, offset)
 
@@ -18,6 +42,15 @@ class Barrel(Tile):
             self.smoke_explosion.render(surf, offset=offset)
 
     def update(self, tilemap, player, enemies):
+        """
+        Update the barrel. It will explode if it is destroyed. The explosion will kill the player and enemies and destroy the tiles around it.
+
+        Parameters:
+            tilemap (Tilemap): The tilemap object where the barrel is.
+            player (Player): The player object.
+            enemies (list[Enemy]): The list of enemies.
+        """
+
         if self.exploded:
             kill = self.smoke_explosion.update()
             if kill:
@@ -40,7 +73,7 @@ class Barrel(Tile):
                     self.pos[1] * tilemap.tile_size,
                 )
             ):
-                if tile.type == "barrel" :
+                if tile.type == "barrel":
                     tile.explode()
                 else:
                     tilemap.remove_tile(
@@ -48,19 +81,32 @@ class Barrel(Tile):
                         self.pos,
                         offgrid=False,
                     )
+            if not self.killed:
+                if rect.colliderect(player.rect):
+                    player.kill(1)
 
-            if rect.colliderect(player.rect):
-                player.kill(1)
-
-            for enemy in enemies.copy():
-                if rect.colliderect(enemy.rect):
-                    enemy.dead = True
+                for enemy in enemies.copy():
+                    if rect.colliderect(enemy.rect):
+                        enemy.dead = True
+                self.killed = True
 
             if done:
                 tilemap.remove_tile(self.pos, self.pos, offgrid=False)
 
-    def explode(self):
+    def explode(self, sfx=None):
+        """
+        Explode the barrel. It will create a smoke explosion that will kill the player and enemies around it.
+
+        Parameters:
+            sfx (pygame.mixer.Sound): The sound effect to play when the barrel explodes. Default is None.
+
+        Returns:
+            bool: If the barrel is exploded or not.
+        """
+
         if not self.exploded:
+            if sfx:
+                sfx.play()
             self.smoke_explosion = Particle(
                 self.smoke_animation,
                 "smoke",
